+++ example/a.ts
@@ -1,17 +1,27 @@
-import { Elysia, t } from '../src'
-import z from 'zod'
+import { Prettify } from 'elysia/types'
+import { Elysia, ElysiaCustomStatusResponse, t } from '../src'
 import { req } from '../test/utils'
+import { Tuple } from '../src/types'
 
-const app = new Elysia()
-	.macro('guestOrUser', {
-		resolve: () => {
-			return {
-				user: null
-			}
+type PickOne<T> = T extends any ? T : never
+
+new Elysia().get(
+	'/test',
+	({ status }) => {
+		return status(200, { key2: 's', id: 2 })
+	},
+	{
+		response: {
+			200: t.Union([
+				t.Object({
+					key2: t.String(),
+					id: t.Literal(2)
+				}),
+				t.Object({
+					key: t.Number(),
+					id: t.Literal(1)
+				})
+			])
 		}
-	})
-	.macro('user', {
-		guestOrUser: true,
-		body: t.String(),
-		resolve: ({ body, status, user }) => {}
-	})
+	}
+)