+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.4.6 - 15 Sep 2025
+Improvement:
+- strictly check for 200 inline status code
+- coerce union status value and return type
+
 # 1.4.5 - 15 Sep 2025
 Improvement:
 - soundness for guard, group